% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_has_appl}
\alias{assert_has_appl}
\title{test the APPL binaries}
\usage{
assert_has_appl()
}
\value{
Will return TRUE if binaries are installed and can be located
 and executed, and FALSE otherwise.
}
\description{
Asserts that the C++ binaries for appl have been compiled successfully
}
\examples{
assert_has_appl()

}
