% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomdp.R
\name{pomdp}
\alias{pomdp}
\title{pomdp}
\usage{
pomdp(T, O, R, GAMMA, initial = NULL, mc.cores = getOption("mc.cores", 1L),
  stdout = FALSE, ...)
}
\arguments{
\item{T}{Transition matrix, dimension n_s x n_s x n_a}

\item{O}{Observation matrix, dimension n_s x n_z x n_a}

\item{R}{Reward matrix, dimension n_s x n_a}

\item{GAMMA}{the discount factor}

\item{initial}{initial belief state, optional, defaults to uniform over states}

\item{mc.cores}{number of cores needed for parallel runs.}

\item{...}{additional arguments to appl SARSOP algorithm, see \code{\link{appl}}.}
}
\value{
optimal value and corresponding policy
}
\description{
pomdp
}
\details{
Dimensions are given as number of states (n_s), number of observed states n_z, number of actions n_a
}

