% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomdp_solve.R
\name{pomdp_solve}
\alias{pomdp_solve}
\title{pomdp_solve}
\usage{
pomdp_solve(transition, observation, utility, discount, belief = rep(1,
  dim(observation)[[1]])/dim(observation)[[1]], verbose = TRUE, ...)
}
\arguments{
\item{transition}{Transition matrix, dimension n_s x n_s x n_a}

\item{observation}{Observation matrix, dimension n_s x n_z x n_a}

\item{utility}{Utility/reward matrix, dimension n_s x n_a}

\item{discount}{the discount factor}

\item{verbose}{logical, should the function include a message with pomdp diagnostics (timings, final precision, end condition)}

\item{...}{additional arguments to appl SARSOP algorithm, see \code{\link{appl}}.}

\item{initial}{initial belief state, optional, defaults to uniform over states}
}
\value{
optimal value and corresponding policy
}
\description{
pomdp_solve
}
\details{
Dimensions are given as number of states (n_s), number of observed states n_z, number of actions n_a
}
\examples{
\dontrun{
## May take > 5s
## Use example code to generate matrices for pomdp problem:
source(system.file("examples/fisheries-ex.R", package = "appl"))
## Run the function:
soln <- pomdp_solve(transition, observation, reward, discount, precision = 10)
}

}

