% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_policyx.R
\name{read_policy}
\alias{read_policy}
\title{read_policy}
\usage{
read_policy(initial, file = "output.policy", fs, Num_fs)
}
\arguments{
\item{initial}{belief to compute dot product with alpha vectors.}

\item{file}{path to file to read in policy}

\item{fs}{the initial fully observable state; Only for the case MOMDP}

\item{Num_fs}{dimension of the fully observable state; Only for the case MOMDP}
}
\value{
a list of the max alpha and associated alpha action, as well as all alpha vectors and associated actions.
}
\description{
read_policy
}

