% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_policyx.R
\name{read_policyx}
\alias{read_policyx}
\title{read_policyx}
\usage{
read_policyx(file = "output.policyx")
}
\arguments{
\item{file}{name of the policyx file to be read.}
}
\value{
a list, first element "vectors" is an n_states x n_vectors array of alpha vectors,
  second element is a numeric vector "action" of length n_vectors whose i'th element indicates
  the action corresponding to the i'th alpha vector (column) in the vectors array.
}
\description{
read a .policyx file created by SARSOP and return alpha vectors and associated actions.
}
\examples{
f <- system.file("extdata", "out.policy", package="sarsop", mustWork = TRUE)
policy <- read_policyx(f)

}
