% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_momdpx.R
\name{write_momdpx}
\alias{write_momdpx}
\title{write_momdpx}
\usage{
write_momdpx(P_full, P_par, O, R, gamma, b_full, b_par = rep(1/dim(O)[1],
  dim(O)[1]), file = "input.pomdpx", digits = 4, digits2 = 10,
  format = "f")
}
\arguments{
\item{P_full}{transition matrix for fully observable states}

\item{P_par}{transition matrix for partially observable states}

\item{O}{observation matrix}

\item{R}{reward}

\item{gamma}{discount factor}

\item{b_full}{initial fully observable state}

\item{b_par}{initial belief of partially observable states}

\item{file}{pomdpx file to create}

\item{digits}{precision to round to before normalizing. Leave at 4 since sarsop seems unable to do more?}

\item{format}{floating point format, because sarsop parser doesn't seem to know scientific notation}

\item{digits}{precision to write solution to. Leave at 10, since normalizing requires additional precision}
}
\description{
write_momdpx
}

