% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writepomdpx_POMDP.R
\name{writepomdpx_POMDP}
\alias{writepomdpx_POMDP}
\title{write_pomdpx}
\usage{
writepomdpx_POMDP(P, O, R, gamma, b)
}
\arguments{
\item{P}{Transition matrix, dimension n_s x n_s x n_a}

\item{O}{Observation matrix, dimension n_s x n_z x n_a}

\item{R}{Reward matrix, dimension n_s x n_a}

\item{gamma}{the discount factor}

\item{b}{belief vector}
}
\description{
write the pompdp xml file
}

