% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_log.R
\name{alphas_from_log}
\alias{alphas_from_log}
\title{alphas_from_log}
\usage{
alphas_from_log(parameters, log_dir = ".", metafile = paste0(log_dir,
  "/meta.csv"))
}
\arguments{
\item{parameters}{a data.frame with the desired parameter values as given in metafile}

\item{log_dir}{path to log directory}

\item{metafile}{path to metafile index, assumed to be meta.csv in log_dir}
}
\value{
a list of alpha vectors matching the query, and a data.frame with the rows of the matching metadata.
}
\description{
alphas_from_log
}

