% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_from_log.R
\name{alphas_from_log}
\alias{alphas_from_log}
\title{alphas_from_log}
\usage{
alphas_from_log(meta, log_dir = ".")
}
\arguments{
\item{meta}{a data frame containing the log metadata
for each set of alpha vectors desired, see
\code{\link{meta_from_log}}}

\item{log_dir}{path to log directory}
}
\value{
a list with a matrix of alpha vectors for each
 entry in the provided metadata (as returned by \code{\link{sarsop}}).
}
\description{
Read alpha vectors from a log file.
}
\examples{
\donttest{ # takes > 5s

source(system.file("examples/fisheries-ex.R", package = "sarsop"))
log = tempfile()
alpha <- sarsop(transition, observation, reward, discount, precision = 10,
                log_dir = log)

}
}
