% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appl.R
\name{pomdpsol}
\alias{SARSOP}
\alias{appl}
\alias{polgraph}
\alias{pomdpconvert}
\alias{pomdpeval}
\alias{pomdpsim}
\alias{pomdpsol}
\title{APPL wrappers}
\usage{
pomdpsol(model, output = tempfile(), precision = 25, timeout = 10)

pomdpeval(args = "--help")

pomdpsim(args = "--help")

pomdpconvert(args = "--help")

polgraph(args = "--help")
}
\description{
Wrappers for the APPL executables. The \code{pomdpsol} function solves a model
file and returns the path to the output policy file.
}
\examples{
setwd(tempdir())
file.copy(system.file("models/example.pomdp", package = "appl"), "example.pomdp")
policy <- pomdpsol("example.pomdp")
readLines(policy)
}

