% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_matrices.R
\name{es_matrices}
\alias{es_matrices}
\title{Matrices for the ecosystem services model}
\usage{
es_matrices(S_0 = 40, V = 90, C = 80, K = 10, E = 1, A = 2,
  Q = 1, discount = 0.95)
}
\arguments{
\item{S_0}{total / maximum number of species possible}

\item{V}{value of the ecosystem services}

\item{C}{cost incurred in protecting the ecosystem}

\item{K}{number of critical species}

\item{E}{the expected number of extinctions per year}

\item{A}{number of possible actions. Default is 2 (all or nothing)}

\item{Q}{cost curvature, default 1. Q > 1 increasingly hard to protect all}

\item{discount}{the discount factor.}
}
\value{
list of  transition matrix, observation matrix, and reward matrix
}
\description{
initialize the transition, observation, and reward matrices
 given a transition function, reward function, and state space
}
\examples{
m <- es_matrices()
}
