% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_from_log.R
\name{f_from_log}
\alias{f_from_log}
\title{f from log}
\usage{
f_from_log(meta)
}
\arguments{
\item{meta}{a data frame containing the log metadata
for each set of alpha vectors desired, see
\code{\link{meta_from_log}}}
}
\value{
the growth function associated with the model indicated.
}
\description{
Read transition function from log
}
\details{
note this function is unique to the fisheries example problem and assumes that
sarsop call is run with logging specifying a column "model" that contains either the string
"ricker" (corresponding to a Ricker-type growth function) or "allen" (corresponding to an Allen-type.)
}
\examples{
\donttest{ # takes > 5s

source(system.file("examples/fisheries-ex.R", package = "sarsop"))
log = tempfile()
alpha <- sarsop(transition, observation, reward, discount, precision = 10,
                log_dir = log)

}
}
