% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.R
\name{fisheries_matrices}
\alias{fisheries_matrices}
\title{fisheries_matrices}
\usage{
fisheries_matrices(states = 0:23, actions = states,
  observed_states = states, reward_fn = function(x, a) pmin(x, a),
  f = ricker, sigma_g = 0.1, sigma_m = sigma_g, type = c("lognormal",
  "uniform"))
}
\arguments{
\item{states}{sequence of possible states}

\item{actions}{sequence of possible actions}

\item{observed_states}{sequence of possible observations}

\item{reward_fn}{function of x and a that gives reward for tacking action a when state is x}

\item{f}{transition function of state x and action a.}

\item{sigma_g}{log-sd for log-normal shock to f or half-width of uniform shock}

\item{sigma_m}{log-sd for log-normal measurement y of state x, or half-width of of uniform}

\item{type}{distribution for noise, "lognormal" or "uniform"}
}
\value{
list of transitition matrix, observation matrix, and reward matrix
}
\description{
initialize the transition, observation, and reward matrices given a transition function, reward function, and state space
}
\details{
assumes log-normally distributed observation errors and process errors
}

