% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_from_log.R
\name{meta_from_log}
\alias{meta_from_log}
\title{meta from log}
\usage{
meta_from_log(
  parameters,
  log_dir = ".",
  metafile = paste0(log_dir, "/meta.csv")
)
}
\arguments{
\item{parameters}{a data.frame with the desired parameter values as given in metafile}

\item{log_dir}{path to log directory}

\item{metafile}{path to metafile index, assumed to be meta.csv in log_dir}
}
\value{
a data.frame with the rows of the matching metadata.
}
\description{
load metadata from a log file
}
\examples{
\donttest{ # takes > 5s

source(system.file("examples/fisheries-ex.R", package = "sarsop"))
log = tempfile()
alpha <- sarsop(transition, observation, reward, discount, precision = 10,
                log_dir = log)

}
}
