% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomdp_solve.R
\name{run_pomdp}
\alias{run_pomdp}
\title{run_pomdp}
\usage{
run_pomdp(transition, observation, utility, discount, initial = rep(1,
  dim(observation)[[1]])/dim(observation)[[1]], verbose = TRUE, ...)
}
\arguments{
\item{transition}{Transition matrix, dimension n_s x n_s x n_a}

\item{observation}{Observation matrix, dimension n_s x n_z x n_a}

\item{utility}{Utility/reward matrix, dimension n_s x n_a}

\item{discount}{the discount factor}

\item{initial}{initial belief state, optional, defaults to uniform over states}

\item{verbose}{logical, should the function include a message with pomdp diagnostics (timings, final precision, end condition)}

\item{...}{additional arguments to appl SARSOP algorithm, see \code{\link{appl}}.}
}
\value{
alpha vectors and corresponding actions
}
\description{
run_pomdp
}

